Refactor :  relocate notes data handler
